#56 fix username already exist exception and invalid login exception

+++ backend/src/main/java/team/YongAndJoe/NewsTodayBackend/entity/User.java
@@ -1,15 +1,17 @@
 package team.YongAndJoe.NewsTodayBackend.entity;
 
 public class User {
-    private int id;
+
+    private Long id;
+
     private String username;
     private String password;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
